Added notifications to update other views after loading new sermon JSON or when searching from the master view controller rather than traversing the view controller hierarchy.
Added sermon update UI notifications for updating table view cells for downloading.

Added a notification for updating the playing / paused button on the top/master view controller when the stream is paused or played from the detail or pushed view controller.

Added didSet for notifications and for clearing the search text.

Replaced didRotateFrom... with viewWillTransition...

Cleaned up search processing to avoid errors.

Added support to tell whether the search button had been clicked before leaving the top view so when you come back the search bar properly resumes first responder - or doesn't.

Restricted removing of notifications to the objects for which they were setup - to avoid clearing others that I want to keep.

Resolved how to properly scroll the collection view after reloading by using dispatch background of a dispatch of the scrolling code on the main thread.

Changed Download to be a class like the other projects and added methods, removed download methods in Sermon.

Changed viewWillTransition in UICollectionViewController class to not reload, which causes recalculating of the icon size if the app isn't active - which it was, by assuming the device was in portrait as it went into the background.

Made the dict strings String? rather than just String, as we did in the later projects.

Added Global variables as proxies for user default lookups for seriesSelected and sermonSelected.